# improve performance of `indexes`