--- a/CHANGES.txt
+++ b/CHANGES.txt
@@ -9,6 +9,8 @@ Drop Python 2 support.
 
 Drop Python 3.5 support and tag Python 3.9 support.
 
+Add ``indexes_optimized`` for faster iteration over bitset members.
+
 
 Version 0.7.16
 --------------