Hi, just tried this neat project and it nicely solved some memory problems I was having with giant sets, however ``indexes`` became a bit of a hot spot. The following is a performance tweak for that function, which seems to be better for both small and large integers:

```python
def indexes(n):
    i = 0
    while n:
        if n & 1:
            yield i
        i += 1
        n >>= 1

def indexes_new(n):
    for i, b in enumerate(bin(n)[-1:1:-1]):
        if b == '1':
            yield i

n = 401247

%%timeit
tuple(indexes(n))
# 3.12 µs ± 31.1 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

%%timeit
tuple(indexes_new(n))
# 1.85 µs ± 16.8 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

n = 49571049247102492471029471029473836150147123059127451024974774770129471029471092470129381410293414012345970273

%%timeit
tuple(indexes(n))
# 57.6 µs ± 890 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%%timeit
tuple(indexes_new(n))
# 23.7 µs ± 112 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Edit - I just check some 'sparse' bits (e.g. `1 << 200 | 1 << 100`) and the story is similar.