@mikulatomas 2021-02-13
Me personally, I love every performance upgrade :) So hope that @xflr6 will have some time to test and itegrate it. Thanks @jcmgray!