@xflr6 2021-02-14
Thanks for the PR (sorry for the delay).

Sounds good (allocating the binary string repr twice might look a bit odd, but whatever works).

Checked that using `bin(n)[2:]` also seems to be faster than other candidates that come to mind such as `f'{n:b}'` or `bin_ = '{0:b}'.format` and `bin_(n)`.

I think I would prefer to keep the current implementation (just for reference): WDYT about add this one as `indexes_optimized` and then use that internally?

Feel free to document the change in `CHANGES.txt`