@jcmgray 2021-02-14
> Sounds good (allocating the binary string repr twice might look a bit odd, but whatever works).

Yeah python performance is not very rational it seems, e.g. my understanding is that calls like `i += 1; n >>= 1` just allocate new ints as well.

> I think I would prefer to keep the current implementation (just for reference): WDYT about add this one as indexes_optimized and then use that internally?

Do you mean just add ``indexes_optimized`` for users internally to call, or make the change here also:

https://github.com/xflr6/bitsets/blob/master/bitsets/bases.py#L23