@xflr6 2021-02-14
> Do you mean just add indexes_optimized for users internally to call, or make the change here also

The second: the current version will only be left for reference (but still doctested) and we use the optimized one in `bases.py`.