@xflr6 2021-02-15
Released [0.8](https://pypi.org/project/bitsets/0.8/#files) (c742889aa53f15f2b691777da5ecee978730729d)